requirements.txt
Flask
tensorflow
keras
numpy
Pillow
fpdf
opencv-python-headless
werkzeug
gunicorn
requests          
groq               
python-dotenv     